Corrected dates in the forecast, added RoomDB to display saved cities

+++ app/src/main/java/com/example/weatherapp/openweathermap/forecast/Current.java
@@ -34,7 +34,7 @@ public class Current {
     public Double dewPoint;
     @SerializedName("uvi")
     @Expose
-    public Integer uvi;
+    public Double uvi;
     @SerializedName("clouds")
     @Expose
     public Integer clouds;
